[TreeMissionExecutor] Remove the initial STEP_INTO command
The execution of the command is asynchronous. When the user already receives the MissionHandle from the Mole, he still cannot use it to instruct commands because the Mole will say that there is a command in progress (even though the user haven't sent any command yet).
This behavior is not compliant with the Mole's docs: At this point in time, commands can be sent to the mission instance through the instruct(MissionHandle, Strand, StrandCommand)} method.